refactor(studio): hoist the catalog labels, the machine sentence and the status dot
The acquisition label, the SIZE/FETCH accounting and the one plain sentence
about a machine were desktop-local while web built its own words for the same
things. They move to studio (and the status dot to ui) so both shells read one
source; desktop re-exports so every existing import keeps working.

Also gives the three output kinds their own "nothing to open onto" sentence in
studio, which fixes web's "3-d object" — the label was lowercased at the call
site instead of using the kind's noun.
