+++ desktop/src/lib/catalog.test.ts
@@ -1,14 +1,5 @@
 import { describe, expect, it } from "vitest";
-import {
-  catalogFetchCaption,
-  catalogIdentityKey,
-  catalogPageUrl,
-  catalogPullLabel,
-  catalogSizeInfo,
-  catalogSizeLabel,
-  isCatalogId,
-  sortInstalledFirst,
-} from "./catalog";
+import { catalogPageUrl, catalogPullLabel, catalogSizeInfo, sortInstalledFirst } from "./catalog";
 import type { CatalogEntry } from "./api/types";
 
 function entry(part: Partial<CatalogEntry>): CatalogEntry {
@@ -24,89 +15,19 @@ function entry(part: Partial<CatalogEntry>): CatalogEntry {
   };
 }
 
-describe("catalogSizeInfo", () => {
-  it("adds shared components to the weights for the fetch total", () => {
+/* The size accounting, the acquisition label and the catalog-id helpers moved
+ * to `@studio/lib/catalogLabel` and are covered by `catalogLabel.test.ts`. This
+ * one case stays: it proves desktop's own wire type still satisfies the shared
+ * structural type through the re-export. */
+describe("the shared catalog labels reach desktop through the re-export", () => {
+  it("labels the acquisition button from a desktop CatalogEntry", () => {
     const info = catalogSizeInfo(
       entry({
         size_bytes: 23_900_000_000,
-        companion_details: [
-          { name: "t5-xxl", size_bytes: 9_200_000_000 },
-          { name: "clip-l", size_bytes: 1_700_000_000 },
-        ],
+        companion_details: [{ name: "t5", size_bytes: 9_200_000_000 }],
       }),
     );
-    expect(info.weightsBytes).toBe(23_900_000_000);
-    expect(info.fetchBytes).toBe(34_800_000_000);
-    expect(info.differs).toBe(true);
-  });
-
-  it("does not diverge when there are no companions", () => {
-    const info = catalogSizeInfo(entry({ size_bytes: 6_400_000_000 }));
-    expect(info.fetchBytes).toBe(6_400_000_000);
-    expect(info.differs).toBe(false);
-  });
-
-  it("copes with an unknown weight size", () => {
-    const info = catalogSizeInfo(entry({ size_bytes: null }));
-    expect(info.weightsBytes).toBeNull();
-    expect(info.fetchBytes).toBeNull();
-    expect(info.differs).toBe(false);
-  });
-});
-
-describe("catalog labels", () => {
-  const two = catalogSizeInfo(
-    entry({
-      size_bytes: 23_900_000_000,
-      companion_details: [{ name: "t5", size_bytes: 9_200_000_000 }],
-    }),
-  );
-  const one = catalogSizeInfo(entry({ size_bytes: 6_400_000_000 }));
-
-  it("shows SIZE · FETCH only when they differ", () => {
-    expect(catalogSizeLabel(two)).toBe("SIZE 23.9 GB · FETCH 33.1 GB");
-    expect(catalogSizeLabel(one)).toBe("SIZE 6.4 GB");
-  });
-
-  it("captions the fetch total only when shared components are added", () => {
-    expect(catalogFetchCaption(two)).toBe("33.1 GB to download, including shared components");
-    expect(catalogFetchCaption(one)).toBeNull();
-  });
-
-  it("labels the Pull button with the download total", () => {
-    expect(catalogPullLabel(two)).toBe("Pull · 33.1 GB");
-    expect(catalogPullLabel(one)).toBe("Pull · 6.4 GB");
-    // Desktop speaks the lexicon; the verb is the caller's.
-    expect(catalogPullLabel(one, "Get it")).toBe("Get it · 6.4 GB");
-  });
-});
-
-describe("isCatalogId", () => {
-  it("recognizes cv: and hf: ids, not plain names", () => {
-    expect(isCatalogId("cv:8001")).toBe(true);
-    expect(isCatalogId("hf:author/model")).toBe(true);
-    expect(isCatalogId("flux-dev:q8")).toBe(false);
-  });
-});
-
-describe("catalogIdentityKey", () => {
-  it("uses source plus a non-empty upstream identity", () => {
-    expect(
-      catalogIdentityKey(entry({ id: "legacy-row", source: "hf", source_id: " author/model " })),
-    ).toBe("hf:author/model");
-  });
-
-  it("does not treat a human name or an empty source id as model identity", () => {
-    expect(
-      catalogIdentityKey(
-        entry({
-          id: "legacy-row",
-          source: "civitai",
-          source_id: " ",
-          name: "Shared human title",
-        }),
-      ),
-    ).toBeNull();
+    expect(catalogPullLabel(info, "Get it")).toBe("Get it · 33.1 GB");
   });
 });
