+++ desktop/src/lib/format.ts
@@ -83,17 +83,12 @@ export function formatEta(seconds: number | null): string {
   return `${Math.floor(minutes / 60)}h ${minutes % 60}m`;
 }
 
-/** Compact uptime for host telemetry: 42 → "42s", 300 → "5m", 5400 → "1h 30m", 200_000 → "2d 7h". */
-export function formatUptime(seconds: number): string {
-  const whole = Math.max(0, Math.floor(seconds));
-  if (whole < 60) return `${whole}s`;
-  const minutes = Math.floor(whole / 60);
-  if (minutes < 60) return `${minutes}m`;
-  const hours = Math.floor(minutes / 60);
-  if (hours < 24) return `${hours}h ${minutes % 60}m`;
-  const days = Math.floor(hours / 24);
-  return `${days}d ${hours % 24}h`;
-}
+/** Compact uptime for host telemetry: 42 → "42s", 300 → "5m", 5400 → "1h 30m", 200_000 → "2d 7h".
+ *
+ * Re-exported from `@studio/lib/formatUptime`, which owns it: the machine
+ * sentence is shared with web and cannot import from a shell. Every existing
+ * caller keeps importing it here. */
+export { formatUptime } from "@studio/lib/formatUptime";
 
 /** Compact relative timestamp for MRU lists ("just now", "5m ago", "3d ago").
  *
